# docs/swagger.yaml
basePath: /tasks
definitions:
  models.Task:
    properties:
      completed:
        description: Completion status of the task
        type: boolean
      id:
        description: ID of the task (string format)
        type: string
      title:
        description: Title of the task
        type: string
    type: object
  models.UpdateReq:
    properties:
      completed:
        description: New completion status for the task update
        type: boolean
      title:
        description: New title for the task update
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Manage Your Tasks
  title: Task Organizator
  version: "1.0"
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Returns the data of all the tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "500":
          description: Internal Server Error
      summary: Get a list of all tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Creates a new task with a unique ID if not provided
      parameters:
      - description: Task object to be created
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new task
      tags:
      - Tasks
  /tasks/:
    delete:
      consumes:
      - application/json
      description: Deletes All Data
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "500":
          description: Internal Server Error
      summary: Deletes All Tasks
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a task with the specified ID
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a task by ID
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Retrieves a task with the specified ID
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Updates a task with the specified ID
      parameters:
      - description: Task ID
        format: int64
        in: path
        name: id
        required: true
        type: string
      - description: Task object with fields to be updated
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateReq'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a task by ID
      tags:
      - Tasks
swagger: "2.0"